+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/lib/EzGradleBomManager.kt
@@ -1,6 +1,5 @@
 package me.thens.ezgradle.lib
 
-import com.github.a7hens.ezgradle.com.github.a7hens.ezgradle.gradle.plugin.BuildConfig
 import me.thens.ezgradle.util.hasConfiguration
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
@@ -10,7 +9,7 @@ class EzGradleBomManager(val project: Project) {
 
     fun addDependencies(version: String) {
         project.dependencies {
-            val bom = platform("${BuildConfig.GROUP}:$EZGRADLE_BOM:$version")
+            val bom = platform("$GROUP:$EZGRADLE_BOM:$version")
             configurations
                 .filter { project.hasConfiguration(it) }
                 .forEach { add(it, bom) }
@@ -18,6 +17,7 @@ class EzGradleBomManager(val project: Project) {
     }
 
     companion object {
+        private const val GROUP: String = "com.github.7hens.ezgradle"
         private const val EZGRADLE_BOM = "ezgradle-bom"
 
         private val configurations = listOf(