+++ ezgradle-run/src/main/java/me/thens/ezgradle/EzGradleRun.kt
@@ -1,11 +1,11 @@
 package me.thens.ezgradle
 
-import me.thens.ezgradle.task.GenerateBuildConfigTask
+import me.thens.ezgradle.task.UpdateBomTask
 
 class EzGradleRun : EzGradleTask {
-    private val tasks = mapOf(
-        "generateBuildConfig" to GenerateBuildConfigTask(),
-    )
+    private val tasks = listOf(
+        UpdateBomTask()
+    ).associateBy { it.taskName }
 
     override fun run(argv: List<String>) {
         val taskName = argv[0]