+++ ezgradle-run/src/main/java/me/thens/ezgradle/EzGradleTask.kt
@@ -1,5 +1,5 @@
 package me.thens.ezgradle
 
-interface EzGradleTask {
+fun interface EzGradleTask {
     fun run(argv: List<String>)
 }