+++ ezgradle-run/src/main/java/me/thens/ezgradle/task/GenerateBuildConfigTask.kt
@@ -1,9 +1,0 @@
-package me.thens.ezgradle.task
-
-import me.thens.ezgradle.EzGradleTask
-
-class GenerateBuildConfigTask: EzGradleTask {
-    override fun run(argv: List<String>) {
-        TODO("Not yet implemented")
-    }
-}